requirements.txt
accessible_output2
configparser
GPUtil
inputs
keyboard
legendary-gl
mss
numpy
Pillow
psutil
pyaudio
pygame
pywin32
requests
scipy
soundfile
tensorboard
termcolor
tqdm
wheel
winshell
numpy==1.24.3